move spec files
