+++ .babelrc
@@ -13,7 +13,7 @@
         "plugins": [
           ["istanbul", {
             "exclude": [
-              "test/spec/**/*.spec.js"
+              "test/**/*.spec.js"
           ]
         }]
       ]
