+++ build-config/fragments/common.js
@@ -5,15 +5,21 @@
 
 const path = require('path');
 const webpack = require('webpack');
+const moment = require('moment');
 
-const time = new Date();
+const time = moment().format('YYYY');
 const pckg = require(path.join(__dirname, '..', '..', 'package.json'));
 
+// add banner with copyright and version info
 var bannerPlugin = new webpack.BannerPlugin(
-`${pckg.name} ${pckg.version} (${time})
-${pckg.homepage}
+`${pckg.name}
+@version ${pckg.version}
+@see ${pckg.homepage}
+@copyright 2014-${time} ${pckg.author}
 @license ${pckg.license}`
 );
+
+// inject version number
 var replaceVersionPlugin = new webpack.DefinePlugin({
   '__VERSION__': JSON.stringify(pckg.version)
 });
@@ -38,7 +44,7 @@ module.exports = {
         rules: [
             {
                 test: /\.js$/,
-                exclude: /(node_modules|bower_components)/,
+                exclude: /(node_modules|bower_components|test)/,
                 use: {
                     loader: 'babel-loader'
                 }
