+++ package.json
@@ -40,7 +40,6 @@
     ".babelrc",
     ".jscsrc",
     "dist/",
-    "es5",
     "src/",
     "test/",
     "scripts/"
@@ -65,7 +64,6 @@
     "babel-loader": "^7.1.4",
     "babel-plugin-add-module-exports": "^0.2.1",
     "babel-plugin-istanbul": "^4.1.5",
-    "babel-plugin-transform-define": "^1.3.0",
     "babel-preset-env": "^1.6.1",
     "babelify": "^8.0.0",
     "browserify": "16",
@@ -96,7 +94,6 @@
     "npm-run-all": "^4.1.2",
     "portscanner": "^2.1.1",
     "qunit": "^2.5.1",
-    "replace": "^0.3.0",
     "serve-static": "^1.13.2",
     "shelljs": "^0.8.1",
     "uglify-js": "3",
