+++ test/utils.spec.js
@@ -1,6 +1,6 @@
-import formatTime from '../../src/js/utils/format-time.js';
+import formatTime from '../src/js/utils/format-time.js';
 
-//import Wavesurfer from '../../src/js/videojs.wavesurfer.js';
+//import Wavesurfer from '../src/js/videojs.wavesurfer.js';
 
 /** @test {format-time} */
 describe('format-time:', function() {
